scripts/stock-analysis/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Financial data
yfinance>=0.2.28

# NLP and sentiment analysis
nltk>=3.8.1

# Database
supabase-py>=2.0.0
postgrest>=0.10.6

# Environment variables
python-dotenv>=1.0.0

# Data visualization (optional, for local testing)
matplotlib>=3.7.0
seaborn>=0.12.0

# Date handling
pytz>=2023.3
